AWS lambda back end javacode for the package
Copied over the second file contained next to the index file in AWS management console. As stated in earlier one, my java is extremely weak since I'm pretty new. I know html and css from class, but Java is next semester and I want to have this available in time for the debates.