Updated the description and added more utterances
Front end interaction model for the Alexa Skill. Name of the intent is linked to a specific policy page (in order as it is on his website) and is the intent AWS lambda uses to handle requests. The samples (utterances) are various ways a user can ask Alexa to fetch that particular intent. Need 15-20 sample utterances before machine learning kicks in and can handle the rest (in theory). More never hurts but less definitely would.